친구 등록, 친구 요청 수락 flow 수정
