+++ src/main/java/com/example/skptemp/domain/notification/util/NotificationUtil.java
@@ -8,7 +8,9 @@ public class NotificationUtil {
 
     public static final String COK_MESSAGE_FORMAT = "%s 님이 \"%s\" 부적을 찔렀어요!";
     public static final String FRIEND_REQUESTED_FORMAT = "요청을 보냈어요.";
+    public static final String FRIEND_REQUESTED_MESSAGE_FORMAT = "%s님이 친구맺기 요청을 보냈어요!";
     public static final String FRIEND_ACCEPTED_FORMAT = "요청을 수락했어요.";
+    public static final String FRIEND_ACCEPTED_MESSAGE_FORMAT = "%s님이 친구맺기 요청을 수락했어요!";
     public static final String CHEER_MESSAGE_FORMAT = "%s 님이 \"%s\" 응원 메세지를 남겼어요. 어서 확인 해보세요!";
     public static final String CHEER_MESSAGE_WITH_ITEM_FORMAT = "%s 님이 \"%s\" 응원 메세지와 함께 선물을 남겼어요. 어서 확인 해보세요!";
 }
