+++ src/main/java/com/example/skptemp/domain/user/dto/FriendCreateRequest.java
@@ -5,4 +5,6 @@ import lombok.Getter;
 @Getter
 public class FriendCreateRequest {
     Long userId;
+    Long notificationId;
+
 }
