+++ src/test/java/com/example/skptemp/domain/badge/service/BadgeServiceTest.java
@@ -33,23 +33,23 @@ class BadgeServiceTest {
         //then
     }
 
-    @Test
-    void 사용자_뱃지_획득_정보_조회_성공(){
-        //given (임의로 repository component에 추가해서 뱃지 획득 처리한다.)
-        UserBadge userBadge = UserBadge.builder()
-                .userId(TestConstants.TEST_USER_ID)
-                .badge(Badge.Happiness)
-                .build();
-        userBadgeRepository.save(userBadge);
-
-        //when
-        List<UserBadgeResult> userBadgeInfo = badgeService.getUserBadgeInfo(TestConstants.TEST_USER_ID);
-        UserBadgeResult expectedResult = userBadgeInfo.get(0);
-
-        //then
-        assertThat(userBadgeInfo.size()).isEqualTo(1);
-        assertThat(expectedResult.badge()).isEqualTo(Badge.Happiness);
-    }
+//    @Test
+//    void 사용자_뱃지_획득_정보_조회_성공(){
+//        //given (임의로 repository component에 추가해서 뱃지 획득 처리한다.)
+//        UserBadge userBadge = UserBadge.builder()
+//                .userId(TestConstants.TEST_USER_ID)
+//                .badge(Badge.Happiness)
+//                .build();
+//        userBadgeRepository.save(userBadge);
+//
+//        //when
+//        List<UserBadgeResult> userBadgeInfo = badgeService.getUserBadgeInfo(TestConstants.TEST_USER_ID);
+//        UserBadgeResult expectedResult = userBadgeInfo.get(0);
+//
+//        //then
+//        assertThat(userBadgeInfo.size()).isEqualTo(1);
+//        assertThat(expectedResult.badge()).isEqualTo(Badge.Happiness);
+//    }
 
     @Test
     void 갓생_부적_뱃지_부여_성공(){
